+++ cedar-policy-core/src/parser/err.rs
@@ -139,12 +139,13 @@ pub enum ToASTErrorKind {
     #[error("a policy with id `{0}` already exists in the policy set")]
     DuplicatePolicyId(ast::PolicyID),
     /// Returned when a template is encountered but a static policy is expected
-    #[error("expected a static policy, got a template containing the slot {}", slot.id)]
-    #[diagnostic(help("try removing the template slot(s) from this policy"))]
-    UnexpectedTemplate {
-        /// Slot that was found (which is not valid in a static policy)
-        slot: ast::Slot,
-    },
+    #[error(transparent)]
+    #[diagnostic(transparent)]
+    ExpectedStaticPolicy(#[from] parse_errors::ExpectedStaticPolicy),
+    /// Returned when a static policy is encountered but a template is expected
+    #[error(transparent)]
+    #[diagnostic(transparent)]
+    ExpectedTemplate(#[from] parse_errors::ExpectedTemplate),
     /// Returned when we attempt to parse a policy or template with duplicate or
     /// conflicting annotations
     #[error("duplicate annotation: @{0}")]
@@ -406,6 +407,16 @@ impl ToASTErrorKind {
         parse_errors::SlotsInConditionClause { slot, clause_type }.into()
     }
 
+    /// Constructor for the [`ToASTErrorKind::ExpectedStaticPolicy`] error
+    pub fn expected_static_policy(slot: ast::Slot) -> Self {
+        parse_errors::ExpectedStaticPolicy { slot }.into()
+    }
+
+    /// Constructor for the [`ToASTErrorKind::ExpectedTemplate`] error
+    pub fn expected_template() -> Self {
+        parse_errors::ExpectedTemplate::new().into()
+    }
+
     /// Constructor for the [`ToASTErrorKind::WrongEntityArgument`] error when
     /// one kind of entity argument was expected
     pub fn wrong_entity_argument_one_expected(
@@ -437,6 +448,41 @@ pub mod parse_errors {
     use std::sync::Arc;
 
     use super::*;
+
+    /// Details about a `ExpectedStaticPolicy` error.
+    #[derive(Debug, Clone, Diagnostic, Error, PartialEq, Eq)]
+    #[error("expected a static policy, got a template containing the slot {}", slot.id)]
+    #[diagnostic(help("try removing the template slot(s) from this policy"))]
+    pub struct ExpectedStaticPolicy {
+        /// Slot that was found (which is not valid in a static policy)
+        pub(crate) slot: ast::Slot,
+    }
+
+    impl From<ast::UnexpectedSlotError> for ExpectedStaticPolicy {
+        fn from(err: ast::UnexpectedSlotError) -> Self {
+            match err {
+                ast::UnexpectedSlotError::FoundSlot(slot) => Self { slot },
+            }
+        }
+    }
+
+    /// Details about a `ExpectedTemplate` error.
+    #[derive(Debug, Clone, Diagnostic, Error, PartialEq, Eq)]
+    #[error("expected a template, got a static policy")]
+    #[diagnostic(help("a template should include slot(s) `?principal` or `?resource`"))]
+    pub struct ExpectedTemplate {
+        /// A private field, just so the public interface notes this as a
+        /// private-fields struct and not a empty-fields struct for semver
+        /// purposes (e.g., consumers cannot construct this type with
+        /// `ExpectedTemplate {}`)
+        _dummy: (),
+    }
+
+    impl ExpectedTemplate {
+        pub(crate) fn new() -> Self {
+            Self { _dummy: () }
+        }
+    }
 
     /// Details about a `SlotsInConditionClause` error.
     #[derive(Debug, Clone, Diagnostic, Error, PartialEq, Eq)]