+++ cedar-policy/CHANGELOG.md
@@ -87,6 +87,9 @@ Cedar Language Version: 4.0
   valid Cedar identifiers. (#1004, resolving #994)
 - (*) Schemas can now reference entity and common types defined in the empty namespace,
   even in contexts occurring in a non-empty namespace. (#1060, resolving #579)
+- `Template` parsing functions (e.g., `Template::parse()`) will now fail when
+  passed a static policy as input. Use the `Policy` parsing functions instead.
+  (#1108, resolving #1095)
 
 ## [3.2.1] - 2024-05-31
 Cedar Language Version: 3.3